Update nuget to 1.0.14 - Fix method Ack(IEnumerable<ulong> deliveryTag) on Subscription class which did the wrong thing by ack all msgs up to the max value of the list, same fix for CompositeSubscription - Add method AckAllUpTo(ulong deliveryTag) on Subscrition to ack all msgs which have delivery tag less than or equal the provided delivery tag - Change the PreFetchSize data type to uint which enventually let the app prefetch everything in the queue if the size is larger than ushort.MaxValue - Add warning log if attempt to declare/bind with Exchange whose name is empty string as that's the default builtin Exchange. Publish msg to that exchange with the routing key equal to the queue name will route the message to that queue

+++ src/Burrow.Extras/Internal/InMemoryPriorityQueue.cs
@@ -19,7 +19,7 @@ namespace Burrow.Extras.Internal
     internal class InMemoryPriorityQueue<T> : IInMemoryPriorityQueue<T> where T : IPriorityMessage
     {
         protected volatile bool Closing;
-        private readonly int _maxSize;
+        private readonly long _maxSize;
         protected IntervalHeap<T> PriorityQueue;
 
 
@@ -43,7 +43,7 @@ namespace Burrow.Extras.Internal
             }
         }
 
-        public InMemoryPriorityQueue(int maxSize, IComparer<T> comparer)
+        public InMemoryPriorityQueue(long maxSize, IComparer<T> comparer)
         {
             _maxSize = maxSize;
             PriorityQueue = new IntervalHeap<T>(comparer);
